requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
jinja2
pillow
pydantic